[TEC-3791] Complete order on checking succeeded captures sum to be >= than order.total.amount, `submitted_at` to be present and order's balance to be <= 0
Separate an `update_order_native_attrs` method in WebhookService for reuse in `capture_upserted_v2` and `order_placed_v2` hooks, because these events could arrive in an undefined order, so we need to check order for completeness in both of these hooks.

Spree::FlowIoOrderDecorator.flow_io_captures_sum method added for returning the sum of amounts of succeeded captures.