[TEC-4780] Refactored SessionsController, CurrentZoneLoader and OrderSync to not deal with Flow session expiration - this logic will be handled by FlowJS on the frontend.
When a visitor hits the site a Flow Session will be created if the country is from a Spree::Zone with Flow Experience. The session_current endpoint will return the flow_session_id to the frontend, and further session refresh will be carried by FlowJS.

When the user will push `Continue to Checkout`, the frontend will pass the current flow_session_id to the `checkout_url` endpoint, the order will be synced to the flow.io authenticated by the received flow_session_id. A request to create a checkout_token will be fired to flow.io, binding the order to the current flow_session_id. The received checkout_token will be used for building the checkout_url, which will be returned to the frontend to redirect the user to flow.io checkout page.

No other order syncs will happen - only before redirecting to checkout.